Анимашки при старте работают не так как надо, с логаутом проблемы, добавил красивые поля ввода, анимации переходов между фрагментами, иконку приложения и русскую локализацию

+++ app/src/main/java/com/example/myapplication/screen/setting/SettingViewModel.kt
@@ -5,7 +5,6 @@ import androidx.lifecycle.asLiveData
 import com.example.myapplication.repository.CloudRepository
 import com.example.myapplication.repository.UsersRepository
 import com.example.myapplication.support.CoroutineViewModel
-import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
 
@@ -14,8 +13,8 @@ class SettingViewModel(
 ) : CoroutineViewModel() {
     val progressLiveDate = MutableLiveData<Boolean>()
 
-    val userNameLiveDate = usersRepository.userName.map { it ?: "" }.asLiveData()
-    fun logout() {
+    val userNameLiveDate = usersRepository.userName.asLiveData()
+    fun logout(){
         launch {
             usersRepository.logout()
         }
@@ -23,9 +22,11 @@ class SettingViewModel(
 
     fun deleteUser() {
         launch {
-            progressLiveDate.postValue(cloudRepository.exportEmptyNotes())
-            usersRepository.deleteUser()
+            val result = cloudRepository.exportEmptyNotes()
+            if (result) {
+                usersRepository.deleteUser()
+            }
+            progressLiveDate.postValue(result)
         }
     }
-
 }
